+++ src/main/java/gtclassic/GTConfig.java
@@ -29,6 +29,7 @@ public class GTConfig {
 	public static boolean morePistonRecipes = true;
 	public static boolean clearerWater = true;
 	public static boolean replaceOceanGravelWithSand = true;
+	public static boolean caveZombiesSpawnWithPickaxe = true;
 	// generation
 	public static boolean bauxiteGenerate = true;
 	public static int bauxiteSize = 16;
@@ -87,6 +88,7 @@ public class GTConfig {
 		morePistonRecipes = cfg.getBoolean("morePistonRecipes", CATEGORY_CONFIG, morePistonRecipes, "Adds more metals to vanilla piston recipe");
 		clearerWater = cfg.getBoolean("clearerWater", CATEGORY_CONFIG, clearerWater, "Reduces fog/haze when underwater");
 		replaceOceanGravelWithSand = cfg.getBoolean("replaceOceanGravelWithSand", CATEGORY_CONFIG, replaceOceanGravelWithSand, "Replaces most gravel in oceans with sand");
+		caveZombiesSpawnWithPickaxe = cfg.getBoolean("caveZombiesSpawnWithPickaxe", CATEGORY_CONFIG, caveZombiesSpawnWithPickaxe, "Zombies spawned in deep caves have a chance of wielding a pickaxe");
 	}
 
 	private static void initGenerationConfig(Configuration cfg) {