+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 # 1.03 (Not Released)
+- Added, Zombie's who spawn underground have a chance to spawn carrying a pickaxe. Ofc with config option.
 - Fixed Baubles items deciding to register whenever they felt like it.
 - Added Neon, and Argon, and centrifuge air separating. Cram it Bear im not adding a whole machine/multi for this, it has no use.
 - Methane and Hydrogen gas blocks now explode if near a fire source, or player walks into them holding a torch.